Decouple encryption, decryption and secret ops
RE: https://github.com/OWASP/Go-SCP/issues/68

This piece of addition helps users to create a custom package out of a simple piece of code/functionality without worrying about going through decoupling process. It provides the barebones so it's up to the user to enhance/refactor etc. while inheriting it. In addition to the decoupled approach, the `nonce` doesn't need to be passed around. It is being generated/extracted within the decoupled methods. The whole example originates from http://www.inanzzz.com/index.php/post/f3pe/data-encryption-and-decryption-with-a-secret-key-in-golang but the string casting has been removed and left to the user to handle if it is needed so the methods are less opinionated on a such minor point.